% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\docType{package}
\name{assign.locations}
\alias{assign.locations}
\alias{assign.locations-package}
\title{\code{assign.locations} package}
\description{
assign.locations
}
\details{
See the Vignette in the future
}
