% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{fun.l.address}
\alias{fun.l.address}
\title{fun.l.address}
\usage{
fun.l.address(pkg.data.root)
}
\arguments{
\item{pkg.data.root}{dropbox root}
}
\description{
Assigns a list of addresses
}
\keyword{clean,}
\keyword{list}
\keyword{locations,}
\keyword{parcels,}
\keyword{update,}
