% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.address.ids}
\alias{funImport.address.ids}
\title{funImport.address.ids}
\usage{
funImport.address.ids(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Creates unique address.ids for places that are different
}
\keyword{address.ids}
