% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.address}
\alias{funImport.address}
\title{funImport.address}
\usage{
funImport.address(x, lookup.address = NULL)
}
\arguments{
\item{x}{data.table}

\item{lookup.address}{data.table used to fill missing zips and cities usually parcel.address}
}
\description{
Explodes and cleans address
}
\keyword{check,}
\keyword{import,}
\keyword{spatial}
