% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.check.list}
\alias{funImport.check.list}
\title{funImport.check.list}
\usage{
funImport.check.list(x)
}
\arguments{
\item{x}{data.table}
}
\description{
checks to see what data fields are being used
}
\keyword{check,}
\keyword{import,}
\keyword{spatial}
