% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.data}
\alias{funImport.data}
\title{funImport.data}
\usage{
funImport.data(l)
}
\arguments{
\item{l}{list of data}
}
\description{
checks to see what data fields are being used
}
\keyword{check,}
\keyword{import,}
\keyword{spatial}
