% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.location.address}
\alias{funImport.location.address}
\title{funImport.location.address}
\usage{
funImport.location.address(l.location, x)
}
\arguments{
\item{l.location}{points, parcels, geocodes.bad, outlines}

\item{x}{the address file (to be assigned to l.locations)}
}
\description{
string matches address (x) to shapefiles address
}
\keyword{assign}
\keyword{import}
\keyword{location}
