% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.location.geocode}
\alias{funImport.location.geocode}
\title{funImport.location.geocode}
\usage{
funImport.location.geocode(locs, address, pkg.data.root, api.key,
  l.study.extent, location.source)
}
\arguments{
\item{locs}{table showing what shapes the address.ids are associated with}

\item{address}{data to be assigned}

\item{pkg.data.root}{source dropbox data}

\item{api.key}{google api key}

\item{l.study.extent}{cities, zips, states}

\item{location.source}{source data mmed.1348998}
}
\description{
string matches address (x) to shapefiles address
}
\keyword{assign}
\keyword{import}
\keyword{location}
