% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.location.intersect}
\alias{funImport.location.intersect}
\title{funImport.location.intersect}
\usage{
funImport.location.intersect(locs, address, pkg.data.root, location.source)
}
\arguments{
\item{locs}{assignment table}

\item{address}{to be looked up}

\item{pkg.data.root}{dropbox data root dir}

\item{location.source}{data source}
}
\description{
intersect points to shapes
}
\keyword{assign}
\keyword{import}
\keyword{location}
