% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport.locations}
\alias{funImport.locations}
\title{funImport.locations}
\usage{
funImport.locations(l, match.threshold, l.study.extent, pkg.data.root)
}
\arguments{
\item{l}{list with all location data.tables}

\item{match.threshold}{string matching criteria}

\item{l.study.extent}{cities, zips, states}

\item{pkg.data.root}{dropbox pkg root}
}
\description{
assigns address to locations (shapefiles)
}
\keyword{assign,}
\keyword{import}
\keyword{location,}
