% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funImport}
\alias{funImport}
\title{funImport}
\usage{
funImport(D, source = "", match.threshold = 0.6, l.study.extent = NULL,
  pkg.data.root = NULL)
}
\arguments{
\item{D}{raw data with all fields}

\item{source}{data source name}

\item{match.threshold}{string match quality}

\item{l.study.extent}{list with cities, zips, states}

\item{pkg.data.root}{path to dropbox pkg.data root}
}
\description{
Assigns address to shapes
}
\keyword{clean}
\keyword{parcels,}
