% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.locations.R
\name{funUpdate.locs}
\alias{funUpdate.locs}
\title{funUpdate.locs}
\usage{
funUpdate.locs(address, match.threshold, pkg.data.root)
}
\arguments{
\item{address}{data.table with all exploded address fields}

\item{match.threshold}{string match quality}

\item{pkg.data.root}{package path data}
}
\description{
Assigns address to locations in last stage of import
}
\keyword{clean}
\keyword{locations}
\keyword{parcels}
\keyword{update}
